requirements_full.txt
# CLM Automation System - Complete Requirements
# Virtual Environment: clm_env

# Core Web Framework
streamlit==1.49.1

# Data Processing
pandas==2.3.2
numpy==2.3.3
python-dateutil==2.9.0.post0

# Document Processing
python-docx==1.2.0
lxml==6.0.1

# Environment & Configuration
python-dotenv==1.1.1

# Logging
loguru==0.7.3

# Machine Learning
scikit-learn==1.7.2
scipy==1.16.2

# AI/ML Framework
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
langchain-core==0.3.76
langchain-text-splitters==0.3.11

# OpenAI Integration
openai==1.108.0
tiktoken==0.11.0

# Vector Database
chromadb==1.1.0

# Additional Dependencies (installed automatically)
# - All required sub-dependencies for the above packages
# - ChromaDB dependencies (onnxruntime, grpcio, etc.)
# - LangChain dependencies (pydantic, sqlalchemy, etc.)
# - Streamlit dependencies (altair, pillow, etc.)



